Prepare for deployment in a docker swarm
