+++ .dockerignore
@@ -1,0 +1,8 @@
+.git/
+.pytest_cache/
+__pycache__/
+
+.dockerignore
+.env*
+!.env.example
+docker-compose.override.yml
