+++ client/Dockerfile.dev
@@ -1,0 +1,4 @@
+FROM node:lts-buster
+WORKDIR /app
+
+USER node
