+++ docker-compose.override.yml.example
@@ -1,0 +1,17 @@
+version: "3"
+
+services:
+  keycloak:
+    image: jboss/keycloak:latest
+    container_name: gg2-keycloak
+    environment:
+      - KEYCLOAK_USER=ggw
+      - KEYCLOAK_PASSWORD=ggw
+      - KEYCLOAK_IMPORT=/tmp/keycloak_ggw.json
+      - DB_VENDOR=postgres
+      - DB_ADDR=postgres
+      - DB_USER=postgres
+      - DB_PASSWORD=abcdef
+    restart: "no"
+    volumes:
+      - ./keycloak_ggw.json:/tmp/keycloak_ggw.json
