--- server/alembic/versions/0a01d5f58771_add_genbank_column_to_vector.py
@@ -1,42 +1,0 @@
-"""Add genbank column to vector
-
-Revision ID: 0a01d5f58771
-Revises: 6dcd6fb67536
-Create Date: 2022-05-16 14:13:41.947168
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = "0a01d5f58771"
-down_revision = "6dcd6fb67536"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table("vector_references", schema=None) as batch_op:
-        batch_op.drop_index("ix_references_id")
-        batch_op.create_index(
-            batch_op.f("ix_vector_references_id"), ["id"], unique=False
-        )
-
-    with op.batch_alter_table("vectors", schema=None) as batch_op:
-        batch_op.add_column(sa.Column("genbank", sa.String()))
-
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table("vectors", schema=None) as batch_op:
-        batch_op.drop_column("genbank")
-
-    with op.batch_alter_table("vector_references", schema=None) as batch_op:
-        batch_op.drop_index(batch_op.f("ix_vector_references_id"))
-        batch_op.create_index("ix_references_id", ["id"], unique=False)
-
-    # ### end Alembic commands ###
