--- server/alembic/versions/33a82a5c6134_changed_vector_type_to_experiment.py
@@ -1,34 +1,0 @@
-"""Changed vector_type to experiment
-
-Revision ID: 33a82a5c6134
-Revises: 0a01d5f58771
-Create Date: 2022-05-30 16:31:39.401026
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = "33a82a5c6134"
-down_revision = "0a01d5f58771"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table("vectors", schema=None) as batch_op:
-        batch_op.add_column(sa.Column("experiment", sa.String(), nullable=True))
-        batch_op.drop_column("vector_type")
-
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table("vectors", schema=None) as batch_op:
-        batch_op.add_column(sa.Column("vector_type", sa.VARCHAR(), nullable=True))
-        batch_op.drop_column("experiment")
-
-    # ### end Alembic commands ###
