--- server/alembic/versions/69252b7f20ca_users_table.py
@@ -1,62 +1,0 @@
-"""Users table
-
-Revision ID: 69252b7f20ca
-Revises: 
-Create Date: 2021-11-08 16:02:47.873255
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = "69252b7f20ca"
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "users",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("iss", sa.String(), nullable=False),
-        sa.Column("sub", sa.String(), nullable=False),
-        sa.Column("created", sa.DateTime(), nullable=False),
-        sa.Column("name", sa.String(), nullable=True),
-        sa.Column("role", sa.String(), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_users_id"), "users", ["id"], unique=False)
-    op.create_index(op.f("ix_users_iss"), "users", ["iss"], unique=False)
-    op.create_index(op.f("ix_users_sub"), "users", ["sub"], unique=False)
-    with op.batch_alter_table("users", schema=None) as batch_op:
-        batch_op.create_unique_constraint("login_id", ["iss", "sub"])
-
-    op.create_table(
-        "providers",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("name", sa.String(), nullable=False),
-        sa.Column("issuer", sa.String(), nullable=False),
-        sa.Column("clientid", sa.String(), nullable=False),
-        sa.Column("secret", sa.String(), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("name"),
-    )
-    op.create_index(op.f("ix_providers_id"), "providers", ["id"], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f("ix_providers_id"), table_name="providers")
-    op.drop_table("providers")
-
-    with op.batch_alter_table("users", schema=None) as batch_op:
-        batch_op.drop_constraint("login_id", type_="unique")
-    op.drop_index(op.f("ix_users_sub"), table_name="users")
-    op.drop_index(op.f("ix_users_iss"), table_name="users")
-    op.drop_index(op.f("ix_users_id"), table_name="users")
-    op.drop_table("users")
-    # ### end Alembic commands ###
