--- server/alembic/versions/6dcd6fb67536_change_mpg_number_to_location.py
@@ -1,39 +1,0 @@
-"""change <mpg_number> to <location>
-
-Revision ID: 6dcd6fb67536
-Revises: 12d6f7ae4742
-Create Date: 2022-02-21 15:49:20.940830
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = "6dcd6fb67536"
-down_revision = "12d6f7ae4742"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table("vectors", schema=None) as batch_op:
-        batch_op.add_column(sa.Column("location", sa.Integer(), nullable=False))
-        batch_op.drop_constraint("lvl_mpg", type_="unique")
-        batch_op.create_unique_constraint("lvl_loc", ["level", "location"])
-        batch_op.drop_column("mpg_number")
-        batch_op.drop_column("BsmB1_site")
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table("vectors", schema=None) as batch_op:
-        batch_op.add_column(sa.Column("mpg_number", sa.INTEGER(), nullable=False))
-        batch_op.drop_constraint("lvl_loc", type_="unique")
-        batch_op.create_unique_constraint("lvl_mpg", ["level", "mpg_number"])
-        batch_op.drop_column("location")
-        batch_op.add_column(sa.Column("BsmB1_site", sa.VARCHAR(), nullable=True))
-
-    # ### end Alembic commands ###
