+++ server/app/config.py
@@ -2,13 +2,13 @@
 
 import secrets
 
-from pydantic import BaseSettings
+from pydantic import BaseSettings, PostgresDsn
 
 
 class Settings(BaseSettings):
     "Global HTTP API server configuration"
 
-    DATABASE_URL: str = "sqlite:///ggw.sqlite?check_same_thread=false"
+    DATABASE_URL: PostgresDsn
 
     API_SECRET: str = secrets.token_urlsafe(32)
     API_JWT_ALGORITHM: str = "HS256"
