+++ server/pyproject.toml
@@ -14,6 +14,8 @@ httpx = "^0.22.0"
 fastapi = "^0.78.0"
 biopython = "^1.79"
 python-multipart = "^0.0.5"
+SQLAlchemy-Utils = "^0.38.2"
+psycopg2-binary = "^2.9.3"
 
 [tool.poetry.dev-dependencies]
 pytest = "^7"
