+++ server/run-dev-env.sh
@@ -1,8 +1,6 @@
 #!/bin/sh
 
-export DATABASE_URL="sqlite:///ggw.sqlite?check_same_thread=false"
 export AUTH_REDIRECT_URI="http://localhost:49999/oidc_login"
-export API_SECRET="THIS IS NOT SECRET"
 
 # Initialise the database
 poetry run alembic upgrade head
@@ -17,7 +15,7 @@ poetry run ggwc add-identity-provider --issuer https://services-staging.vib.be/ 
 poetry run ggwc create-admin --iss 1 --sub 4af3d9fb-beb3-4b59-af33-d325addbf1bb ggw
 
 # Populate the database
-poetry run ggwc import ./Test_data/Import_File_Test.csv  ./Test_data/Example_Genbank_Files/ 4af3d9fb-beb3-4b59-af33-d325addbf1bb
+poetry run ggwc import ./Test_data/Import_File_Test.csv ./Test_data/Example_Genbank_Files/ 4af3d9fb-beb3-4b59-af33-d325addbf1bb
 
 # Run the app
-poetry run hypercorn --bind "0.0.0.0:8000" main:app --reload
+poetry run hypercorn --root-path /api/v1 --bind "0.0.0.0:8000" app.main:app --reload
